face_identification/cosface/requirements.txt
numpy==1.23.4
onnx==1.12.0
Pillow==9.3.0
torchvision==0.14.0
